feat(refactor): 전체 피드백 내용 적용
**
강사님의 피드백을 반영하여 JCF(Java Collections Framework)라는
   구체적인 구현 기술이 명시되었지만 그 이상의 기능을 가져 표준 구현체 클래스 이름인 Impl로 변경합니다

멘토님의 피드백을 반영하여 기존 정적 팩토리 메서드에서 protected -> private 로 접근 제한을 수정합니다

NPE예방을 위해 ChannelMessageRepositoryImpl의 findAllById 메서드의 equals비교 순서를 역순으로 변경합니다.

nullcheck없는 부분에 nullcheck를 추가합니다

Optional의 isPresent 메서드를 성능 문제로 orElseThrow || orElse || ifPresentOrElse 메서드로 변경하였습니다.

기존 AppConfig를 AppConfigRegacy로 변경하고 더 이상 사용 되지 않습니다.
**

+++ src/main/java/com/sprint/mission/discodeit/repository/impl/ChannelMessageRepositoryImpl.java
@@ -1,26 +1,33 @@
-package com.sprint.mission.discodeit.repository.jcf;
+package com.sprint.mission.discodeit.repository.impl;
 
 import com.sprint.mission.discodeit.entity.ChannelMessage;
 import com.sprint.mission.discodeit.repository.ChannelMessageRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
-public class JCFChannelMessageRepository extends JCFBaseRepository<ChannelMessage, UUID> implements ChannelMessageRepository {
+@Repository
+public class ChannelMessageRepositoryImpl extends BaseRepositoryImpl<ChannelMessage, UUID> implements ChannelMessageRepository {
 
     @Override
     public List<ChannelMessage> findAllByChannelId(UUID channelId) {
+        if (dataMap.isEmpty()) {
+            return List.of();
+        }
         return dataMap.values().stream()
-                .filter(cm -> !cm.isDeleted() && cm.getChannelId().equals(channelId))
+                .filter(cm -> !cm.isDeleted() && channelId.equals(cm.getChannelId()))
                 .sorted(Comparator.comparing(ChannelMessage::getCreatedAt)) // 생성 시간(오름차순)으로 정렬
                 .toList();
     }
 
     @Override
     public List<ChannelMessage> findAllBySenderId(UUID senderId) {
-        return dataMap.values().stream()
+        if(dataMap.isEmpty()){
+            return List.of();
+        }
+        return  dataMap.values().stream()
                 .filter(cm -> !cm.isDeleted() && cm.getSenderId().equals(senderId))
                 .sorted(Comparator.comparing(ChannelMessage::getCreatedAt))
                 .toList();
@@ -28,9 +35,10 @@ public class JCFChannelMessageRepository extends JCFBaseRepository<ChannelMessag
 
     @Override
     public void deleteAllBySenderId(UUID senderId) {
+        if(findAllBySenderId(senderId).isEmpty()){
+            return;
+        }
         findAllBySenderId(senderId).stream()
                 .map(ChannelMessage::getId).forEach(this::deleteById);
     }
-
-
 }
